master.cfg: Remove "make refs-default"
"make refs-default" polls www.kernel.org, and the polls were failing
because the Python library uses a boring default for User-Agent.

But more than that, "make refs-default" is something you should do
after altering the static refs YAML file, and then you commit the
files generated by that target. Not something that needs to be done
on every CI run.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>